Refactor survey question and modify for in report builder
