+++ front/app/containers/Admin/projects/project/nativeSurvey/FormResults/FormResultsQuestion/MultipleChoice/CompletionBar.tsx
