+++ front/app/containers/Admin/projects/project/nativeSurvey/FormResults/FormResultsQuestion/TextQuestion/Analysis/ConsentModal.tsx
@@ -10,7 +10,7 @@ import {
   Checkbox,
 } from '@citizenlab/cl2-component-library';
 import { useIntl, FormattedMessage } from 'utils/cl-intl';
-import messages from '../messages';
+import messages from '../../../messages';
 
 import Divider from 'components/admin/Divider';
