+++ front/app/containers/Admin/projects/project/nativeSurvey/FormResults/FormResultsQuestion/TextQuestion/Analysis/CreateAnalysisModal.tsx
@@ -8,7 +8,7 @@ import {
   Checkbox,
 } from '@citizenlab/cl2-component-library';
 import { useIntl, FormattedMessage } from 'utils/cl-intl';
-import messages from '../messages';
+import messages from '../../../messages';
 import { useParams, useSearchParams } from 'react-router-dom';
 
 import useAddAnalysis from 'api/analyses/useAddAnalysis';
