+++ front/app/containers/Admin/projects/project/nativeSurvey/FormResults/FormResultsQuestion/TextQuestion/Analysis/index.tsx
@@ -9,7 +9,7 @@ import {
   colors,
 } from '@citizenlab/cl2-component-library';
 import { useIntl } from 'utils/cl-intl';
-import messages from '../messages';
+import messages from '../../../messages';
 
 import Modal from 'components/UI/Modal';
