+++ front/app/containers/Admin/projects/project/nativeSurvey/FormResults/index.tsx
@@ -4,22 +4,18 @@ import { useParams } from 'react-router-dom';
 
 // Hooks
 import useLocale from 'hooks/useLocale';
+import useFormResults from 'api/survey_results/useSurveyResults';
+import useProjectById from 'api/projects/useProjectById';
 
 // components
 import { Box, Text } from '@citizenlab/cl2-component-library';
+import FormResultsQuestion from './FormResultsQuestion';
 
 // i18n
 import messages from '../messages';
 
 // utils
 import { isNilOrError } from 'utils/helperUtils';
-
-// hooks
-import useFormResults from 'api/survey_results/useSurveyResults';
-import useProjectById from 'api/projects/useProjectById';
-
-// Services
-import FormResultsQuestion from './FormResultsQuestion';
 
 const FormResults = () => {
   const { projectId, phaseId } = useParams() as {
