+++ front/app/containers/Admin/reporting/components/ReportBuilder/Widgets/SurveyResultsWidget/SurveyResults.tsx
@@ -9,7 +9,7 @@ import { useSurveyResults } from 'api/graph_data_units';
 // components
 import { Box, Text } from '@citizenlab/cl2-component-library';
 import NoData from '../_shared/NoData';
-import FormResultsQuestion from 'containers/Admin/projects/project/nativeSurvey/FormResults/FormResultsQuestion';
+import FormResultsQuestion from './FormResultsQuestion';
 import Dot from './Dot';
 import PageBreakBox from 'components/admin/ContentBuilder/Widgets/PageBreakBox';
 
@@ -72,7 +72,7 @@ const SurveyResults = ({ phaseId, shownQuestions }: Props) => {
   return (
     <>
       <Box width="100%" mb="24px">
-        <Text variant="bodyM" color="primary" mt="0px" mb="0px">
+        <Text variant="bodyM" mt="0px" mb="0px">
           {'| '}
           {localize(project.data.attributes.title_multiloc)}
           {phase && (
